requirements.txt
google-auth
google-auth-oauthlib
google-generative-ai